requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==1.7.1
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
